fix: moved internal utils to separate file to avoid poluting edge runtime
